Optimize perf of file list population
The list code was calling ListView_SetItem for every added item in the
list.  This doesn't seem necessary, and takes a lot of time. Everything
seems to continue to work correctly without this, and it makes showing
the list with a lot of items up to 3x faster in my measurements.