Add 'Select Projects...' option to filter dropdown
https://github.com/galenelias/gotofile-vs/issues/6

Current project drop down lets you either search all projects or one
specific project.  There is a feature request to instead allow selecting
a subset of projects instead.

The feature request had much grander designs on the feature, but as time
is limited, this is just a bare bones implementation of a simple dialog
that lets you select the projects to filter, with some persistence.