Add 'CharBag' first pass filter optimization
Adds a 'CharBag' class which keeps a bit mask of which characters are in
a string to allow for efficient superset checks to ensure each filepath
contains the required characters for each filter string.

For 2+ character searches, this is roughly a ~3x speedups in lookup
speed. Not really very beneficial in practice, but it's pretty
straightforward, so why not!